bot/requirements.txt
qdrant-client
torch
sentence-transformers
python-telegram-bot